_archive/web_dashboard_pre_v13.5/requirements.txt
# Core (required)
fastapi
uvicorn[standard]
websockets

# Optional (for Docker service integration)
psycopg2-binary  # PostgreSQL database
httpx             # RAG API proxy
neo4j             # Attack graph queries
